+++ src/Guilded.Base/events/CalendarRsvpManyEvent.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.ExceptionServices;
-using System.Threading.Tasks;
 using Guilded.Base.Content;
 using Guilded.Base.Servers;
 using Newtonsoft.Json;