+++ src/Guilded.Commands/CommandConfiguration.cs
@@ -1,17 +1,12 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reactive.Linq;
-using System.Threading.Tasks;
 using Guilded.Base;
-using Guilded.Base.Events;
 
 namespace Guilded.Commands;
 
 /// <summary>
 /// Represents the module that adds <see cref="CommandAttribute">commands</see> to <see cref="BaseGuildedClient">Guilded clients</see>.
 /// </summary>
-/// <seealso cref="CommandBase" />
+/// <seealso cref="CommandParent" />
 /// <seealso cref="CommandAttribute" />
 /// <seealso cref="CommandFallbackAttribute" />
 public class CommandConfiguration