+++ src/Guilded.Commands/CommandModule.cs
@@ -11,10 +11,10 @@ namespace Guilded.Commands;
 /// <summary>
 /// Represents the module that adds <see cref="CommandAttribute">commands</see> to <see cref="BaseGuildedClient">Guilded clients</see>.
 /// </summary>
-/// <seealso cref="CommandBase" />
+/// <seealso cref="CommandParent" />
 /// <seealso cref="CommandAttribute" />
 /// <seealso cref="CommandFallbackAttribute" />
-public class CommandModule : CommandBase
+public abstract class CommandModule : CommandParent
 {
     #region Fields
     private AbstractGuildedClient? _subscribedClient;